move scalar aggregate benchmarks back to compute/kernels
